WIP: sort of working but should be completed
Important: got a problem to install vmlua.def. I disabled it but should
find a way to re-enable it.

Now we find the architecture with lj_arch_test so the previous code is
redundant and should be disabled.

The target_arch_flags are determined by lj_arch_test but not currently
used. The same for dasm_arch.

Cosmetic: in lj_arch_test.c define a struct to hold the array to
const char * and the corresponding array count.